Actualización completa: corrección de payload en Chatwoot, mejoras de logging y sincronización del proyecto

+++ models/user_memory.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, Integer, String, DateTime, INT, Text
+from sqlalchemy import Table, Column, Integer, String, DateTime, BigInteger, Text
 from sqlalchemy import func 
 from config.db import engine, meta_data
 
@@ -7,7 +7,7 @@ user_memory = Table(
     "user_memory",
     meta_data,
     Column("id", Integer, primary_key=True),
-    Column("phone_number", INT, nullable=False),
+    Column("phone_number", BigInteger, nullable=False),
     Column("summary", Text, nullable=False),
     Column("category", String(50), nullable=False),
     Column("created_at", DateTime, nullable=False, server_default=func.now()),
